Script to annotate sunflower PCB with GRID and SEG silkscreens
A script to mark the back of the sunflower PCB with the
LED legs that need to be left uncut because they connect
GRID and SEG lines to the driver board.

This also marks the extra pads that have been added to
the sunflower board where there is no existing LED leg
that will work. In these cases, the pads can end up being
marked with both a circle and a square silkscreen.

Signed-off-by: Andy Bennett <andyjpb@ashurst.eu.org>